+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v1.2.1
+
+*2025-03-11*
+
+**Changed**
+
+- Upgrade the [WikiLint](https://npmjs.com/package/wikilint) package to [v2.17.2](https://github.com/bhsd-harry/wikiparser-node/blob/main/CHANGELOG.md#v1172)
+
 ## v1.2.0
 
 *2025-03-08*